+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 - [BREAKING] Langchain::Vectorsearch::Milvus was rewritten to work with newer milvus 0.10.0 gem
+- [BREAKING] Removing Langchain::LLM::GooglePalm
 - Assistant can now process image_urls in the messages (currently only for OpenAI and Mistral AI)
 - Vectorsearch providers utilize the global Langchain.logger
 - Update required milvus, qdrant and weaviate versions